+++ src/components/baseline.js
@@ -1,35 +1,174 @@
 import React from 'react'
 import {css, createGlobalStyle} from 'styled-components'
-import {Helmet} from 'react-helmet'
+import PropTypes from 'prop-types'
 
+import ThemeProvider from './theme-provider'
 import {fontFamily} from './typography'
 
-export const baselineCss = css`
-  margin: 0;
-  padding: 0;
+const meyerResetCss = css`
+  /* http://meyerweb.com/eric/tools/css/reset/ 
+v2.0 | 20110126
+License: none (public domain)
+*/
 
-  box-sizing: border-box;
-  font-family: ${fontFamily};
-
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-
-  fill: transparent;
+  html,
+  body,
+  div,
+  span,
+  applet,
+  object,
+  iframe,
+  h1,
+  h2,
+  h3,
+  h4,
+  h5,
+  h6,
+  p,
+  blockquote,
+  pre,
+  a,
+  abbr,
+  acronym,
+  address,
+  big,
+  cite,
+  code,
+  del,
+  dfn,
+  em,
+  img,
+  ins,
+  kbd,
+  q,
+  s,
+  samp,
+  small,
+  strike,
+  strong,
+  sub,
+  sup,
+  tt,
+  var,
+  b,
+  u,
+  i,
+  center,
+  dl,
+  dt,
+  dd,
+  ol,
+  ul,
+  li,
+  fieldset,
+  form,
+  label,
+  legend,
+  table,
+  caption,
+  tbody,
+  tfoot,
+  thead,
+  tr,
+  th,
+  td,
+  article,
+  aside,
+  canvas,
+  details,
+  embed,
+  figure,
+  figcaption,
+  footer,
+  header,
+  hgroup,
+  menu,
+  nav,
+  output,
+  ruby,
+  section,
+  summary,
+  time,
+  mark,
+  audio,
+  video {
+    margin: 0;
+    padding: 0;
+    border: 0;
+    font-size: 100%;
+    vertical-align: baseline;
+  }
+  /* HTML5 display-role reset for older browsers */
+  article,
+  aside,
+  details,
+  figcaption,
+  figure,
+  footer,
+  header,
+  hgroup,
+  menu,
+  nav,
+  section {
+    display: block;
+  }
+  body {
+    line-height: 1;
+  }
+  ol,
+  ul {
+    list-style: none;
+  }
+  blockquote,
+  q {
+    quotes: none;
+  }
+  blockquote:before,
+  blockquote:after,
+  q:before,
+  q:after {
+    content: '';
+    content: none;
+  }
+  table {
+    border-collapse: collapse;
+    border-spacing: 0;
+  }
 `
 
-const BaselineCSS = createGlobalStyle`
-    html, body {
-      ${baselineCss}
-    }
+const andiveResetCss = css`
+  * {
+    box-sizing: border-box;
+    font-family: ${fontFamily};
+
+    -moz-osx-font-smoothing: grayscale;
+    -webkit-font-smoothing: antialiased;
+  }
+
+  a {
+    text-decoration: none;
+  }
+
+  svg {
+    fill: transparent;
+  }
 `
 
-export default function Baseline() {
+const BaselineCss = createGlobalStyle`
+  ${meyerResetCss}
+  ${andiveResetCss}
+`
+
+export default function AndiveProvider({theme, children}) {
   return (
     <>
-      <Helmet>
-        <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Asap" />
-      </Helmet>
-      <BaselineCSS />
+      <BaselineCss />
+      <ThemeProvider theme={theme}>{children}</ThemeProvider>
     </>
   )
+}
+
+AndiveProvider.propTypes = {
+  theme: PropTypes.object,
+  children: PropTypes.node
 }